requirements.txt
git+https://bitbucket.org/50onred/fifty.git@1.2.5#egg=Fifty-Flask
git+https://bitbucket.org/50onred/fifty-tables.git@1.7.0#egg=Fifty-Tables
Flask-Assets==0.11
Flask-Cache==0.13.1
Flask-DebugToolbar==0.10.0
Flask-Login==0.3.2
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-WTF==0.12
Flask-Bcrypt==0.7.1
Flask-Migrate==1.8.0
Jinja2==2.8
MarkupSafe==0.23
Pygments==2.1.3
SQLAlchemy==1.0.13
Sphinx==1.4.4
WTForms==2.1
Werkzeug==0.11.10
blinker==1.4
coverage==4.1
cssmin==0.2.0
docutils==0.12
flake8==2.5.5
itsdangerous==0.24
jsmin==2.2.1
mccabe==0.5.0
pep8==1.7.0
py==1.4.31
pyflakes==1.2.3
pytest==2.9.2
webassets==0.11.1
wsgiref==0.1.2
MySQL-Python==1.2.5
boto==2.40.0
requests==2.10.0
uwsgi==2.0.13.1

